# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Revel Demo API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/journeys/runs/{runId}": {
      "get": {
        "summary": "Get the status of a specific journey run",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Current state of the run",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JourneyRunStatusResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g., non-UUID runId)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationError" }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "error": { "type": "string" } } }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "error": { "type": "string" } } }
              }
            }
          }
        }
      }
    }
...
  },
  "components": {
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "path": { "type": "string" }
              }
            }
          }
        },
        "required": ["error"]
      },
...
      "JourneyRunStatusResponse": {
        "type": "object",
        "properties": {
          "runId": { "type": "string" },
          "status": { "type": "string", "enum": ["in_progress", "completed", "failed"] },
          "journeyId": { "type": "string" },
          "currentNodeId": { "type": "string" },
          "patientContext": { "type": "object", "additionalProperties": true }
        },
        "required": ["runId", "status", "journeyId", "currentNodeId", "patientContext"]
      },
...
    }
  }
}